+++ oak/src/lib.rs
@@ -1,3 +1,9 @@
+pub mod acorn;
+pub mod dataset;
+pub mod fvecs;
+pub mod predicate;
+pub mod stubs;
+
 #[cxx::bridge(namespace = "faiss")]
 pub mod ffi {
     unsafe extern "C++" {
@@ -6,7 +12,7 @@ pub mod ffi {
         type IndexACORNFlat;
 
         fn new_index_acorn(
-            d: i32,
+            d: i32, // dimensionality of the vectors being added
             M: i32,
             gamma: i32,
             M_beta: i32,
@@ -21,17 +27,12 @@ pub mod ffi {
 
         unsafe fn search_index(
             idx: &mut UniquePtr<IndexACORNFlat>,
-            n: i64,                       // number of query vectors
-            x: *const f32,                // pointer to an array of the query vectors
-            k: i64,                       // number of vectors to return for each query vector
+            n: i64,                     // number of query vectors
+            x: *const f32,              // pointer to an array of the query vectors
+            k: i64,                     // number of vectors to return for each query vector
             distances: *mut f32, // pointer to an array of (k*n) floats, each representing a distance of the result from the query vector
             labels: *mut i64, // pointer to an array of (k*n) indices, each representing the ID of the query vector in idx
-            filter_id_map: *const c_char, // a bitmap of the IDs in the filter, an array of (n * N) bools, where N is the total number of vectors in the index, and a '1' represents that the vector at that index passes the predicate for that query.
+            filter_id_map: *mut c_char, // a bitmap of the IDs in the filter, an array of (n * N) bools, where N is the total number of vectors in the index, and a '1' represents that the vector at that index passes the predicate for that query.
         );
     }
 }
-
-pub mod architecture;
-pub mod dataset;
-pub mod query;
-pub mod stubs;
